Comments resolved.  Small refactor to move jwt creation to JWTUtils file. Move resetPasswordEmail content to BE ResetPassword
